+++ docs/CHANGELOG.rst
@@ -78,7 +78,8 @@ New Features:
   similar to :func:`~klibs.KLUserInferface.key_pressed` for the keyboard. 
 * Added a new function :func:`~klibs.KLUserInferface.get_clicks` to easily
   fetch the (x, y) coordinates of any mouse clicks in a given input event queue.
-* Added a :meth:`~klibs.KLDatabase.Database.select` method to the 
+* Added :meth:`~klibs.KLDatabase.Database.select` and
+  :meth:`~klibs.KLDatabase.Database.delete` methods to the 
   :class:`~klibs.KLDatabase.Database` class.